+++ pages/[username].tsx
@@ -21,7 +21,7 @@ import GridRep from '~components/GridRep'
 import GridRepCollection from '~components/GridRepCollection'
 import ErrorSection from '~components/ErrorSection'
 import FilterBar from '~components/FilterBar'
-import ProfileHead from '~components/ProfileHead'
+import ProfileHead from '~components/head/ProfileHead'
 
 import type { AxiosError } from 'axios'
 import type { NextApiRequest, NextApiResponse } from 'next'
